pyarrow data keep their original type if possible

+++ pandas/tests/extension/test_arrow.py
@@ -3562,5 +3562,5 @@ def test_cast_dictionary_different_value_dtype(arrow_type):
 def test_map_numeric_na_action():
     ser = pd.Series([32, 40, None], dtype="int64[pyarrow]")
     result = ser.map(lambda x: 42, na_action="ignore")
-    expected = pd.Series([42.0, 42.0, np.nan], dtype="float64")
+    expected = pd.Series([42.0, 42.0, np.nan], dtype=ser.dtype)
     tm.assert_series_equal(result, expected)
